Move down `walk()` and friends from `Utility` to `Basics`, and adjust unit tests.  Comment cleanup and name normalization to follow.
