+++ Tests/Utility/PathTests.swift
@@ -91,81 +91,6 @@ class PathTests: XCTestCase {
     ]
 }
 
-class WalkTests: XCTestCase {
-
-    func testNonRecursive() {
-        var expected: [AbsolutePath] = [ "/usr", "/bin", "/sbin" ]
-
-        for x in walk("/", recursively: false) {
-            if let i = expected.index(of: x) {
-                expected.remove(at: i)
-            }
-            XCTAssertEqual(2, x.components.count)
-        }
-
-        XCTAssertEqual(expected.count, 0)
-    }
-
-    func testRecursive() {
-        let root = AbsolutePath(#file).appending("../../../Sources")
-        var expected: [AbsolutePath] = [ root.appending("Build"), root.appending("Utility") ]
-
-        for x in walk(root) {
-            if let i = expected.index(of: x) {
-                expected.remove(at: i)
-            }
-        }
-
-        XCTAssertEqual(expected.count, 0)
-    }
-
-    func testSymlinksNotWalked() {
-        mktmpdir { root in
-            let root = resolveSymlinks(root)  // FIXME: it would be better to not need this, but we end up relying on /tmp -> /private/tmp.
-            
-            try makeDirectories(root.appending("foo"))
-            try makeDirectories(root.appending("bar/baz/goo"))
-            try symlink(create: root.appending("foo/symlink").asString, pointingAt: root.appending("bar").asString, relativeTo: root.asString)
-
-            XCTAssertTrue(root.appending("foo/symlink").asString.isSymlink)
-            XCTAssertEqual(resolveSymlinks(root.appending("foo/symlink")), root.appending("bar"))
-            XCTAssertTrue(resolveSymlinks(root.appending("foo/symlink/baz")).asString.isDirectory)
-
-            let results = walk(root.appending("foo")).map{ $0 }
-
-            XCTAssertEqual(results, [root.appending("foo/symlink")])
-        }
-    }
-
-    func testWalkingADirectorySymlinkResolvesOnce() {
-        mktmpdir { root in
-            try makeDirectories(root.appending("foo/bar"))
-            try makeDirectories(root.appending("abc/bar"))
-            try symlink(create: root.appending("symlink").asString, pointingAt: root.appending("foo").asString, relativeTo: root.asString)
-            try symlink(create: root.appending("foo/baz").asString, pointingAt: root.appending("abc").asString, relativeTo: root.asString)
-
-            XCTAssertTrue(root.appending("symlink").asString.isSymlink)
-
-            let results = walk(root.appending("symlink")).map{ $0 }.sorted()
-
-            // we recurse a symlink to a directory, so this should work,
-            // but `abc` should not show because `baz` is a symlink too
-            // and that should *not* be followed
-
-            XCTAssertEqual(results, [root.appending("symlink/bar"), root.appending("symlink/baz")])
-        }
-    }
-}
-
-extension WalkTests {
-    static var allTests = [
-        ("testNonRecursive", testNonRecursive),
-        ("testRecursive", testRecursive),
-        ("testSymlinksNotWalked", testSymlinksNotWalked),
-        ("testWalkingADirectorySymlinkResolvesOnce", testWalkingADirectorySymlinkResolvesOnce),
-    ]
-}
-
 class StatTests: XCTestCase {
 
     func test_isdir() {
